CO: a fixed price shouldnt be treated as a discount
